refactor(wrap-native-flow): narrow down WrapUnwrapContext
